info: switch to using UniversalSpringCell
This is a (long overdue) patch to update the info cell to use
UniversalSpringCell.  Most of the wiring is now handled in a spring XML
file.

The patch tries to strike a balance between being complete and trying
to do too much, so that the patch is review-able.  For example,
messages that request serialised data (from web-server) are handled via
the CellMessageReceiver interface but the DataGatheringActivities send
requests for data and processing the responses through a custom
framework.

Target: trunk
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/5689/
Acked-by: Gerd Behrmann
Acked-by: Albert Rossi