cleaner: Add hsqldb support
Addresses the issue that the cleaner was not functional when used with
hsqldb, and thus didn't work in system-test. The trigger to populate
the trash table was missing.

This patch adds a trigger. HSQLDB unfortunately does not support data
manipulation statements in BEFORE DELETE triggers and thus cannot
populate the trash on delete from t_inodes (AFTER DELETE is too late
as the delete cascades to t_locationinfo before the trigger is run,
and thus the data is already gone). Instead the patch adds an
AFTER DELETE trigger on the t_locationinfo table itself. A trigger
condition is used to limit the trigger to deletion of an inode and
thus excludes regular deletes of replicas on pool and tapes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5735/