Fix CellStub timeout handling
A recent change to CellStub introduced a time unit property. Thus
the interpretation of getTimeout() changed, as the value was no
longer guaranteed to be in milliseconds. This patch updates the
code that uses getTimeout() and either lets it use getTimeoutInMillis()
instead, or updates it to be time unit aware. The patch also fixes
several calls to setTimeout().

Thanks to Dmitry for finding this.

Target: trunk
Require-book: no
Require-notes: no
Acked-by: Albert Rossi <arossi@fnal.gov>
Patch: http://rb.dcache.org/dashboard/