pool: Add callout to health check script
dCache pools periodically perform simple I/O tests to verify
that a pool is still working. We can however not check the
general health status of the host system, eg. whether a disk
in the array has failed.

This patch adds the ability to call an external health check
program. If the program exits with a non-zero exit code, the
pool health is considered degraded and will cause the pool to
be marked read-only or disabled.

The command is configurable through the pool.check-health-command
property.

Target: trunk
Require-book: yes
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/5775/