webdav: Fix problems with root paths
Add support for enforcing a user-specific root directory in the WebDAV door.
Attempts to read or write to a directory outside the user's root directory
will result in a permission denied error being reported back to the user.

As a specicial case, if the user has a user-specific root and that user
attempts to read '/' then the reply is a redirection to that user's
home directory (if specified) or their root directory if not.

The patch also updates CSS and adds the missing favicon.

Require-book: yes
Require-notes: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5672/