Fix building of dcache and srmclient RPMs
When building RPM packages of dCache and srm-client, the intent was
that the packages are built within the build directory and the
resulting files end up in the dist directory.  To achieve this,
a complicated hack was employed where the rpmbuild rc file is
loaded from a custom file, which references a new macro to
override existing macro values.

At some point, RPM support for rc files was broken and this
method no longer worked.   New RPM files were written in the
default location (~/rpmbuild/RPMS/...).  Depending on the
build environment, this continued to work, but only "by accident."

This patch removes the fragile rc and macro file approach and
replaces it with direct injection of the values via the command
line.  The --define argument allows overriding of these values
and provides a more robust approach to controlling RPM.

Many properties defined in rpm.properties were simply not used.
These have been removed.

The dist directory was referred to in relative terms.  Since the
RPM command changes directory, this isn't safe.  Therefore the
definition has been updated to use the absolute path.