Fix HsmFlushController Seconds to Milliseconds conversion
HsmFlushController converts the time units for _flushingInterval and
_retryDelayOnError from Seconds to Milliseconds in the setter. When
writing the setup it does not convert them back, but keeps them in
Milliseconds which leads to delays increasing by factor 1000 every time
the setup is saved and the pool restarted.

This patch adds the conversion from Milliseconds to seconds on printing
the setup.

Acked-by: Gerd
Target: master
Require-book: no
Require-notes: no