pool: Fix problems with restoring permanent migration jobs
Addresses a regression introduced in 2.6 in how migration jobs are
restored on pool startup. The issue was that restored migration jobs
would have "null" as the command string, which would show up in
migration ls and migration info output, as well as stored back to
the pool setup file when the save command was used. In the latter
case, a subsequent pool restart would fail.

Target: trunk
Request: 2.6
Require-notes: yes
Require-book: no