Fix names and locations of packaged files.
The built RPM files are currently stored in the area that is copied
into the RPM.  While this works, it means that subsequent builds of
RPMs will include the already built RPMs.  The tar-ball package has the
wrong name and isn't copied into the expected directory.  The Debian
build isn't in the correct location.

These problems are fixed with this patch.