srm: Add failure counters for SRMv2 interface and refine counter formatting
The srm (and many other services) collects stats of requests and failures.
However for the SRMv2 interface failures were not counted. This is rectified
by this patch.

The patch also adjusts the formatting of the request gauges as seen in
the admin shell: The sampling period is printed in seconds rather than
milliseconds (the high resolution is not needed, IMO, and it makes it
hard to read), thousand separators are added to several columns, and a
couple of spelling errors are fixed.

While at it, I removed some dead code in the SRM.

Target: trunk
Require-book: yes
Require-notes: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5895/