Fix more deprecated properties
When creating a new pool, a wait-for-files configuration is added. This generated
wait-for-files uses the deprecated property 'path'.  If the admin configures the pool's path
with the new property ('pool.path') then '${pool}' expands to the empty string and the pool
doesn't start.

This patch fixes this by using the new property.

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/5897/
Acked-by: Tigran Mkrtchyan