doors: fix support for binding to a specific interface
There are a number of problems with how doors handle configuration that describes to which
network interface they should bind, if any.  First, the default value (the empty string) is
not accepted by xrootd and webdav seems to interpret this as "bind to localhost".  Second, the
admin interface (a.k.a. "ssh2") doesn't support binding at all and always listens on all
interfaces.

This patch fixes these problems.  It also changes the place-holder value for wildcard from the
empty string to 'any'.

FOR RELEASE NOTES:

The keyword 'any' is used when specifying a door should listen on any interface (previously it
was the empty string).  The default value has been adjusted accordingly.

Requires-notes: yes
Requires-book: yes
Target: master
Patch: http://rb.dcache.org/r/5898/
Acked-by: Gerd Behrmann