SRM: fix ping
SRM includes code that checks if a reponse is successful.  It does
this by invoking the getReturnStatus that Axis creates, via
reflection.

Unfortunately, not all response objects contain this method (the
response from a ping operation is the only example).  The current
code that evaluates whether a response is successful results in a
NoSuchMethodException being thrown for such operations.  This
exception is treated as a failure, resulting in ping no longer
working.

This patch fixes this problem by catching the NoSuchMethodException
and treating such operations as always succeeding.  (Catching the
exception like this is not nice, but seems the least ugly way of
discovering if the response object contains the desired method).

Target: master
Requires-notes: no
Requires-book: no
Patch: http://rb.dcache.org/r/5904/
Acked-by: Gerd Behrmann