spacemanager: Delete file when cancelling a space reservation
Addresses a problem that may cause inconsistencies between
the files stored in a link group and in spaces of the link
group.

Files are preregistered with space manager before uploading
the file. This reservation can be cancelled before the upload
completes. Such a file will have been authorized for a
particular link group, thus cancelling the reservation should
also break the upload and remove the file.

This patch achieves this by deleting the name space entry
when such a non-final reservation (ie the reservation is
still in RESERVED or TRANSFERRING) is cancelled. This may
happen when the reservation is explicitly cancelled by the
SRM (eg when the client calls srmAbortFiles or the upload
is killed by an srmRm on the SURL), or when the reservation
expires. The file is only deleted if a PNFSID has been
registered, which is the case iff the transfer was authorized
for a link group. The file is deleted by PNFSID, which
guarantees that we do not accidentally delete another file
by the same path.

A consequence of this patch is that if the client does not
finish the upload with a call to srmPutDone, the srm request
will eventually time out and the file will be deleted.

Target: trunk
Request: 2.7
Require-notes: yes
Require-book: yes
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/5942/