x509: fix DN extraction from certificate chain
The x509 gPlazma2 plugin used the JGlobus-2 principal that is
non-backwards compatible with the JGlobus-1.8 GlobusPrincipal. All
other dCache components were updated to use the dCache-supplied
backwards compatible GlobusPrincipal except the x509 plugin.  The
result was that while SRM worked (as it supplies the DN as a correct
GlobusPrincipal), logging in via GridFTP failed (as it supplies only
the certificate chain).

The gPlazma1 plugin is not affect by this.

This patch fixes this by using the correct GlobusPrincipal

FOR RELEASE NOTES:

Fixes a problem with the x509 plugin that prevents users logging into
the GridFTP door if the 'gplazma1' plugin isn't being used.

Request: 2.2
Requires-notes: yes
Requires-book: no
Patch: http://rb.dcache.org/r/5939/
Acked-by: Karsten Schwank