ldap: fail if no mapping, make plugin less noisy
The ldap plugin provides map, session and identity functionality.  As
a map and session plugin, the ldap plugin currently always succeeds.
This prevents an admin from making the ldap plugin requisite.

The plugin also logs at error level.  If the admin configures the
plugin as optional then some failures are expected.  For this reason,
the patch moves the logging down the debug level.

FOR RELEASE NOTES:

Currently, the ldap gPlazma plugin never fails in the map- or session-
phases of a login.  With this change, the plugin will fail if no
username is supplied or if there is no mapping for this user.  Sites
should check their gPlazma configuration if they use the ldap plugin.

Target: master
Request: 2.7
Requires-notes: yes
Requires-book: yes
Patch: http://rb.dcache.org/r/5937
Acked-by: Tigran Mkrtchyan